+++ dropwizard/src/main/scala/fr/davit/akka/http/metrics/dropwizard/DropwizardRegistry.scala
@@ -33,21 +33,21 @@ object DropwizardRegistry {
 class DropwizardRegistry(settings: HttpMetricsSettings)(implicit val underlying: MetricRegistry)
     extends HttpMetricsRegistry(settings) {
 
-  override lazy val active: Gauge = new DropwizardGauge("akka.http.requests.active")
+  override lazy val active: Gauge = new DropwizardGauge(settings.namespace, "requests.active")
 
-  override lazy val requests: Counter = new DropwizardCounter("akka.http.requests")
+  override lazy val requests: Counter = new DropwizardCounter(settings.namespace, "requests")
 
-  override lazy val receivedBytes: Histogram = new DropwizardHistogram("akka.http.requests.bytes")
+  override lazy val receivedBytes: Histogram = new DropwizardHistogram(settings.namespace, "requests.bytes")
 
-  override lazy val responses: Counter = new DropwizardCounter("akka.http.responses")
+  override lazy val responses: Counter = new DropwizardCounter(settings.namespace, "responses")
 
-  override lazy val errors: Counter = new DropwizardCounter("akka.http.responses.errors")
+  override lazy val errors: Counter = new DropwizardCounter(settings.namespace, "responses.errors")
 
-  override lazy val duration: Timer = new DropwizardTimer("akka.http.responses.duration")
+  override lazy val duration: Timer = new DropwizardTimer(settings.namespace, "responses.duration")
 
-  override lazy val sentBytes: Histogram = new DropwizardHistogram("akka.http.responses.bytes")
+  override lazy val sentBytes: Histogram = new DropwizardHistogram(settings.namespace, "responses.bytes")
 
-  override lazy val connected: Gauge = new DropwizardGauge("akka.http.connections.active")
+  override lazy val connected: Gauge = new DropwizardGauge(settings.namespace, "connections.active")
 
-  override lazy val connections: Counter = new DropwizardCounter("akka.http.connections")
+  override lazy val connections: Counter = new DropwizardCounter(settings.namespace, "connections")
 }